refactor: AI 호출과 트랜잭션 분리
